pyproject.toml
[project]
name = "fsg-fin-indicator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.2.2",
    "jaydebeapi>=1.2.3",
    "kuzu>=0.10.0",
    "marimo",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pydantic-ai>=0.2.4",
    "pydantic-ai-slim[mcp]>=0.2.4",
    "pyhive>=0.7.0",
    "pymysql",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.41",
    "thrift>=0.21.0",
    "tqdm>=4.67.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = [
  "."
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
